transcription-tool/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai-whisper
requests>=2.31.0
python-dotenv>=1.0.0

# Additional dependencies for Whisper
torch
torchaudio
numpy<2.0.0
ffmpeg-python
